experience/5713-validate-endpoint: create new base class RequestFunction for ReportFunction and ValidateFunction
