+++ prime-router/src/test/kotlin/azure/ReportFunctionTests.kt
@@ -1,5 +1,6 @@
 package gov.cdc.prime.router.azure
 
+import azure.RequestFunction
 import com.microsoft.azure.functions.HttpStatus
 import gov.cdc.prime.router.ActionLog
 import gov.cdc.prime.router.CovidSender
@@ -66,23 +67,6 @@ class ReportFunctionTests {
         "NO,NO,h8jev96rc,YES,YES,YES,257628001,60001007\n" +
         "abbott,P,95209-3,SARS-CoV+SARS-CoV-2 (COVID-19) Ag [Presence] in Respiratory specimen by Rapid " +
         "immunoassay,419984006,95209-3,202112181841-0500,202112151325-0500,LumiraDx SARS-CoV-2 Ag Test_LumiraDx " +
-        "UK Ltd.,LumiraDx SARS-CoV-2 Ag Test*,SomeEntityID,SomeEntityID,3,2131-1,2135-2,F,19931,Sussex,1404270765," +
-        "Reicherts,NormanB,19931,97D0667471,Any lab USA,DE,19931,122554006,esyuj9,vhd3cfvvt,DE,NO,bgq0b2e," +
-        "840533007,NO,NO,h8jev96rc,YES,YES,YES,257628001,60001007"
-    val csvString_2Records_diff = "senderId,processingModeCode,testOrdered,testName,testResult,testPerformed," +
-        "testResultDate,testReportDate,deviceIdentifier,deviceName,specimenId,testId,patientAge,patientRace," +
-        "patientEthnicity,patientSex,patientZip,patientCounty,orderingProviderNpi,orderingProviderLname," +
-        "orderingProviderFname,orderingProviderZip,performingFacility,performingFacilityName," +
-        "performingFacilityState,performingFacilityZip,specimenSource,patientUniqueId,patientUniqueIdHash," +
-        "patientState,firstTest,previousTestType,previousTestResult,healthcareEmployee,symptomatic,symptomsList," +
-        "hospitalized,symptomsIcu,congregateResident,congregateResidentType,pregnant\n" +
-        "abbottt,P,95209-3,SARS-CoV+SARS-CoV-2 (COVID-19) Ag [Presence] in Respiratory specimen by Rapid " +
-        "immunoassay,419984006,95209-3,202112181841-0500,202112151325-0500,LumiraDx SARS-CoV-2 Ag Test_LumiraDx " +
-        "UK Ltd.,LumiraDx SARS-CoV-2 Ag Test*,SomeEntityID,SomeEntityID,3,2131-1,2135-2,F,19931,Sussex,1404270765," +
-        "ReichertA,NormanBA,19931,97D0667471,Any lab USA,DE,19931,122554006,esyuj9,vhd3cfvvt,DE,NO,bgq0b2e,840533007," +
-        "NO,NO,h8jev96rc,YES,YES,YES,257628001,60001007\n" +
-        "abbott,P,95209-3,SARS-CoV+SARS-CoV-2 (COVID-19) Ag [Presence] in Respiratory specimen by Rapid " +
-        "immunoassayy,419984006,95209-3,202112181841-0500,202112151325-0500,LumiraDx SARS-CoV-2 Ag Test_LumiraDx " +
         "UK Ltd.,LumiraDx SARS-CoV-2 Ag Test*,SomeEntityID,SomeEntityID,3,2131-1,2135-2,F,19931,Sussex,1404270765," +
         "Reicherts,NormanB,19931,97D0667471,Any lab USA,DE,19931,122554006,esyuj9,vhd3cfvvt,DE,NO,bgq0b2e," +
         "840533007,NO,NO,h8jev96rc,YES,YES,YES,257628001,60001007"
@@ -320,7 +304,7 @@ class ReportFunctionTests {
         )
 
         // Invoke function run
-        var res = reportFunc.run(req)
+        val res = reportFunc.run(req)
 
         // verify
         assert(res.statusCode == 400)
@@ -377,7 +361,7 @@ class ReportFunctionTests {
 
         val req = MockHttpRequestMessage(csvString_2Records)
 
-        every { reportFunc.validateRequest(any()) } returns ReportFunction.ValidatedRequest(
+        every { reportFunc.validateRequest(any()) } returns RequestFunction.ValidatedRequest(
             csvString_2Records,
             sender = sender
         )
@@ -396,7 +380,7 @@ class ReportFunctionTests {
         every { accessSpy.isDuplicateItem(any(), any()) } returns true
 
         // act
-        var resp = reportFunc.processRequest(req, sender)
+        val resp = reportFunc.processRequest(req, sender)
 
         // assert
         verify(exactly = 2) { engine.isDuplicateItem(any()) }
